refactor: extract method to get push notification name into extension method
